refactor(types): merge `types.go` into existing files to reduce fragmentation and improve organization
