+++ helmtools/git/tag.go
@@ -13,6 +13,13 @@ import (
 	"github.com/rancher/ob-charts-tool/helmtools/util"
 )
 
+// Tag represents a git tag with its associated metadata.
+type Tag struct {
+	Name       string // Short tag name (e.g., "v1.0.0")
+	Ref        string // Full reference (e.g., "refs/tags/v1.0.0")
+	CommitHash string // Commit hash the tag points to
+}
+
 // VerifyTagExists checks if a tag exists in a remote repository.
 // Returns (exists bool, ref string, hash string, error).
 // The context can be used for cancellation and timeouts.
